Refactor: Member 패키지 구조 수정, Member와 Oauth 통합
