+++ src/main/java/org/retriever/server/dailypet/domain/member/controller/MemberController.java
@@ -1,4 +1,4 @@
-package org.retriever.server.dailypet.domain.member;
+package org.retriever.server.dailypet.domain.member.controller;
 
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.RequestMapping;
