+++ src/main/java/org/retriever/server/dailypet/domain/member/dto/request/KakaoLoginRequestDto.java
@@ -1,4 +1,4 @@
-package org.retriever.server.dailypet.domain.oauth;
+package org.retriever.server.dailypet.domain.member.dto.request;
 
 import lombok.Getter;
